requirements.txt
streamlit==1.28.2
streamlit-authenticator==0.2.3
bcrypt==4.1.2
pandas==2.1.4
pillow==10.1.0
python-dotenv==1.0.0
